initial quaternion

+++ include/constants.h
@@ -19,6 +19,6 @@ constexpr size_t log_file_size = 1; // preallocated size in Mb
 // MEASUREMENTS: pressure, ax, ay, az, gx, gy, gz, mx, my, mz
 constexpr size_t ukf_number_of_states =  15; // number of elements in UKF state vector
 constexpr size_t ukf_number_of_measurements =  10; // number of elements in UKF measurement vector
-constexpr float alpha_coefficient = 1;
-constexpr float beta_coefficient = 2;
-constexpr float kappa_coefficient = 0;
+constexpr double alpha_coefficient = 1;
+constexpr double beta_coefficient = 2;
+constexpr double kappa_coefficient = 0;
